gère les démarches étranges
