+++ packages/ui/src/components/titre.tsx
@@ -17,19 +17,16 @@ import { caminoFiltres } from 'camino-common/src/filters'
 import { ReferencesTypes } from 'camino-common/src/static/referencesTypes'
 import { TableSortEvent } from './_ui/table'
 import { activitesColonneIdAnnee } from './activites'
-import { CaminoDate, dateFormat } from 'camino-common/src/date'
+import { dateFormat } from 'camino-common/src/date'
 import { Alert } from './_ui/alert'
-import { titreIdOrSlugValidator, TitreIdOrSlug, TitreGet, getMostRecentValidValueProp, TitreGetDemarche, TitreId } from 'camino-common/src/titres'
+import { titreIdOrSlugValidator, TitreIdOrSlug, TitreGet, getMostRecentValidValueProp, TitreId } from 'camino-common/src/titres'
 import { TitresLinkForm } from './titre/titres-link-form'
 import { canReadTitreActivites } from 'camino-common/src/permissions/activites'
 import { TitreTimeline } from './titre/titre-timeline'
 import { CommuneId } from 'camino-common/src/static/communes'
 import { canDeleteTitre, canEditTitre, canHaveActivites } from 'camino-common/src/permissions/titres'
 import { TitreDemarche } from '@/components/titre/titre-demarche'
-import { DemarchesStatutsIds } from 'camino-common/src/static/demarchesStatuts'
-import { isEtapeStatusOk } from 'camino-common/src/static/etapesStatuts'
-import { isEtapeDecision } from 'camino-common/src/static/etapesTypes'
-import { isNotNullNorUndefined, isNullOrUndefined, isNullOrUndefinedOrEmpty, OmitDistributive, onlyUnique } from 'camino-common/src/typescript-tools'
+import { isNotNullNorUndefined, isNullOrUndefined, onlyUnique } from 'camino-common/src/typescript-tools'
 import { DsfrIcon } from './_ui/icon'
 import { DsfrSeparator } from './_ui/dsfr-separator'
 import { TitreAbonnerButton } from './titre/titre-abonner-button'
@@ -39,83 +36,11 @@ import { AdministrationId } from 'camino-common/src/static/administrations'
 import { getAdministrationsLocales } from 'camino-common/src/administrations'
 import { getGestionnairesByTitreTypeId } from 'camino-common/src/static/administrationsTitresTypes'
 import { DemarcheEditPopup } from './titre/demarche-edit-popup'
+import { PhaseWithAlterations, phaseWithAlterations } from './titre/phase'
 
 const activitesSort: TableSortEvent = {
   colonne: activitesColonneIdAnnee,
   ordre: 'desc',
-}
-
-export type TitreTimelineEvents = TitreGetDemarche & { first_etape_date: CaminoDate | null }
-export type PhaseWithDateDebut = OmitDistributive<TitreGetDemarche, 'demarche_date_debut'> & { demarche_date_debut: CaminoDate; events: TitreTimelineEvents[] }
-
-export type DemarcheAlteration = TitreGetDemarche & { date_etape_decision_ok: CaminoDate; events: TitreTimelineEvents[] }
-
-export type PhaseWithAlterations = [PhaseWithDateDebut, ...DemarcheAlteration[]][] | [[OmitDistributive<TitreGetDemarche, 'demarche_date_debut'> & { demarche_date_debut: null }]]
-const phaseWiltAlterations = (demarches: TitreGetDemarche[]): PhaseWithAlterations => {
-  if (isNullOrUndefinedOrEmpty(demarches)) {
-    return []
-  }
-  const simplePhases: PhaseWithDateDebut[] = demarches
-    .filter((demarche): demarche is PhaseWithDateDebut => isNotNullNorUndefined(demarche.demarche_date_debut))
-    .map(phase => ({ ...phase, events: [] }))
-
-  if (isNullOrUndefinedOrEmpty(simplePhases)) {
-    if (demarches.length > 1) {
-      throw new Error('Le titre a plusieurs démarches sans phases')
-    }
-
-    return [[{ ...demarches[0], demarche_date_debut: null }]]
-  }
-
-  const phasesWithAlterations: [PhaseWithDateDebut, ...DemarcheAlteration[]][] = simplePhases.map(phase => {
-    const demarchesAlterationsForThisPhase = demarches
-      .map(demarche => {
-        if (isNullOrUndefined(demarche.demarche_date_debut) && [DemarchesStatutsIds.Accepte, DemarchesStatutsIds.Termine].includes(demarche.demarche_statut_id)) {
-          const etapeFound = [...demarche.etapes].reverse().find(etape => isEtapeDecision(etape.etape_type_id) && isEtapeStatusOk(etape.etape_statut_id))
-
-          if (isNotNullNorUndefined(etapeFound) && etapeFound.date > phase.demarche_date_debut && (isNullOrUndefined(phase.demarche_date_fin) || etapeFound.date <= phase.demarche_date_fin)) {
-            return { ...demarche, date_etape_decision_ok: etapeFound.date }
-          }
-        }
-
-        return null
-      })
-      .filter(isNotNullNorUndefined)
-
-    return [phase, ...demarchesAlterationsForThisPhase.map(demarche => ({ ...demarche, events: [] }))]
-  })
-
-  demarches.forEach(demarche => {
-    if (isNotNullNorUndefined(demarche.etapes) && ![DemarchesStatutsIds.Accepte, DemarchesStatutsIds.Termine].includes(demarche.demarche_statut_id)) {
-      const first_etape_date = demarche.etapes.length > 0 ? demarche.etapes[0].date : null
-      if (demarche.demarche_date_debut === null) {
-        if (first_etape_date === null) {
-          const latestPhaseWithAlterations = phasesWithAlterations[phasesWithAlterations.length - 1]
-          const latestAlterations = latestPhaseWithAlterations[latestPhaseWithAlterations.length - 1]
-          latestAlterations.events.push({ ...demarche, first_etape_date: null })
-        } else {
-          const phaseWithAlterations: [PhaseWithDateDebut, ...DemarcheAlteration[]] =
-            phasesWithAlterations.find(p => p[0].demarche_date_debut < first_etape_date && (isNullOrUndefined(p[0].demarche_date_fin) || p[0].demarche_date_fin > first_etape_date)) ??
-            phasesWithAlterations[0]
-
-          const found = [...phaseWithAlterations].reverse().find(stuff => {
-            return (
-              ('date_etape_decision_ok' in stuff && stuff.date_etape_decision_ok < first_etape_date) ||
-              ('demarche_date_debut' in stuff && isNotNullNorUndefined(stuff.demarche_date_debut) && stuff.demarche_date_debut < first_etape_date)
-            )
-          })
-          if (isNotNullNorUndefined(found)) {
-            found.events.push({ ...demarche, first_etape_date })
-          } else {
-            console.error('impossibruh', first_etape_date)
-            console.error('impossibruh', phasesWithAlterations)
-          }
-        }
-      }
-    }
-  })
-
-  return phasesWithAlterations
 }
 
 export const Titre = defineComponent(() => {
@@ -321,7 +246,7 @@ export const PureTitre = defineComponent<Props>(props => {
 
   const phases = computed<PhaseWithAlterations>(() => {
     if (titreData.value.status === 'LOADED') {
-      return phaseWiltAlterations(titreData.value.value.demarches)
+      return phaseWithAlterations(titreData.value.value.demarches)
     }
 
     return []
